deploy/backend-agent/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Azure AI dependencies
azure-identity>=1.19.0
azure-ai-projects==1.0.0b10
azure-ai-evaluation>=1.0.0
azure-ai-agents>=1.0.0b1

# Database
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Async and utilities
httpx>=0.27.2
aiohttp>=3.11.11
python_dotenv>=1.0.1
pydantic==2.10.1
pandas>=2.2.3
pillow>=11.1.0
nest-asyncio>=1.5.0
aiosqlite>=0.20.0